requirements.unstable.txt
beautifulsoup
boto
django<1.6
django-analytical>=0.12.1
django-annoying
django_compressor
django-heroku-memcacheify
django-heroku-postgresify
django-extensions
django-nose
django-selector
django-staticfiles
django-storages
django-ses
fabric
gunicorn
lxml
psycopg2
south
splinter
versiontools